Fix imprecise specular

+++ Packages/com.deltation.toon-shader/Assets/DELTation/ToonShader/Shaders/ToonShaderUtils.hlsl
@@ -79,14 +79,14 @@ float get_aniso_specular(const float3 view_direction_ws, const float3 tangent_ws
     return max(0, specular);
 }
 
-inline half get_specular(half3 view_direction_ws, half3 normal_ws, half3 light_direction_ws)
+inline half get_specular(float3 view_direction_ws, float3 normal_ws, float3 light_direction_ws)
 {
     const half3 half_vector = normalize(view_direction_ws + light_direction_ws);
     return saturate(dot(normal_ws, half_vector));
 }
 
-inline half3 get_specular_color(half3 light_color, half3 view_direction_ws, half3 normal_ws, half3 tangent_ws,
-                                half3 light_direction_ws)
+inline half3 get_specular_color(half3 light_color, float3 view_direction_ws, float3 normal_ws, float3 tangent_ws,
+                                float3 light_direction_ws)
 {
     #ifndef _SPECULAR
     return 0;
@@ -109,12 +109,12 @@ inline half3 get_specular_color(half3 light_color, half3 view_direction_ws, half
     #endif
 }
 
-inline half get_fresnel(half3 view_direction_ws, half3 normal_ws)
+inline half get_fresnel(float3 view_direction_ws, float3 normal_ws)
 {
     return 1 - saturate(dot(view_direction_ws, normal_ws));
 }
 
-inline half3 get_fresnel_color(half3 light_color, half3 view_direction_ws, half3 normal_ws, half brightness)
+inline half3 get_fresnel_color(float3 light_color, float3 view_direction_ws, float3 normal_ws, half brightness)
 {
     #ifndef _FRESNEL
     return 0;
